Continue tweaking display
Annoyingly, we cannot use icons and vbars on the same screen
due to LCDproc or hardware limitations.  Boo.  Fake it with
A - Z.

Redo the door layout to be a checked checkbox if none open, or
an unchecked checkbox with either L(eft), R(ight), or B(oth) shown.

While here, add a heartbeat of sorts to count loops through the
monitor.  If this digit doesn't increment, something's gone wrong!